+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Graph generation as SVG
 - Internal designer allowing elements removal
+- Scripts restitution from designers
 
 ### Changed